forth commit

+++ src/main/java/ro/fasttrackit/project/model/entity/Rating.java
@@ -45,7 +45,7 @@ public class Rating {
 		return movieRating;
 	}
 
-	public void setMovieRating(int movieRating) {
+	public void setMovieRating(double movieRating) {
 		this.movieRating = movieRating;
 	}
